upgrade rltk

+++ src/spawner.rs
@@ -110,7 +110,7 @@ fn goblin(ecs: &mut World, x: i32, y: i32) {
     monster(ecs, x, y, rltk::to_cp437('g'), "Goblin");
 }
 
-fn monster<S: ToString>(ecs: &mut World, x: i32, y: i32, glyph: u8, name: S) {
+fn monster<S: ToString>(ecs: &mut World, x: i32, y: i32, glyph: u16, name: S) {
     ecs.create_entity()
         .with(Position { x, y })
         .with(Renderable {
